+++ frontend/src/components/prompts/ShareInfo.vue
@@ -17,10 +17,10 @@
           <strong>{{ $t("prompts.displayName") }}</strong> {{ req.name }}
         </div>
         <div class="share-info-element" :title="modTime">
-          <strong>{{ $t("prompts.lastModified", { suffix: ":" }) }}</strong> {{ humanTime }}
+          <strong>{{ $t("files.lastModified", { suffix: ":" }) }}</strong> {{ humanTime }}
         </div>
         <div class="share-info-element">
-          <strong>{{ $t("prompts.size", { suffix: ":" }) }}</strong> {{ humanSize }}
+          <strong>{{ $t("general.size", { suffix: ":" }) }}</strong> {{ humanSize }}
         </div>
       </div>